+++ mini_template/src/renderer.rs
@@ -1,21 +1,27 @@
-use crate::Template;
+use crate::{Template, TemplateKey};
 use std::collections::HashMap;
 
 use crate::value::VariableManager;
 
 use super::modifier::Modifier;
 
-pub struct RenderContext<'a, VM: VariableManager> {
+pub struct RenderContext<'a, VM: VariableManager, TK = String>
+where
+    TK: TemplateKey,
+{
     pub modifier: &'a HashMap<&'static str, &'a Modifier>,
     pub variables: VM,
-    pub templates: &'a HashMap<u64, Template>,
+    pub templates: &'a HashMap<TK, Template>,
 }
 
-impl<'a, VM: VariableManager> RenderContext<'a, VM> {
+impl<'a, 'b, VM: VariableManager, TK> RenderContext<'a, VM, TK>
+where
+    TK: TemplateKey,
+{
     pub fn new(
         modifier: &'a HashMap<&'static str, &'a Modifier>,
         variables: VM,
-        templates: &'a HashMap<u64, Template>,
+        templates: &'a HashMap<TK, Template>,
     ) -> Self {
         Self {
             modifier,
@@ -50,7 +56,7 @@ mod tests {
         let tpl = parse(tpl).unwrap();
         let mut rendered = String::new();
         tpl.render(
-            &mut RenderContext::new(&HashMap::new(), HashMap::new(), &HashMap::new()),
+            &mut RenderContext::<_, String>::new(&HashMap::new(), HashMap::new(), &HashMap::new()),
             &mut rendered,
         )
         .unwrap();
@@ -66,7 +72,7 @@ mod tests {
         let mut rendered = String::new();
 
         tpl.render(
-            &mut RenderContext::new(&HashMap::new(), variables, &HashMap::new()),
+            &mut RenderContext::<_, String>::new(&HashMap::new(), variables, &HashMap::new()),
             &mut rendered,
         )
         .unwrap();
@@ -88,7 +94,7 @@ mod tests {
         let mut rendered = String::new();
 
         tpl.render(
-            &mut RenderContext::new(&HashMap::new(), variables, &HashMap::new()),
+            &mut RenderContext::<_, String>::new(&HashMap::new(), variables, &HashMap::new()),
             &mut rendered,
         )
         .unwrap();
@@ -121,7 +127,7 @@ mod tests {
         let mut rendered = String::new();
 
         tpl.render(
-            &mut RenderContext::new(&modifiers, variables, &HashMap::new()),
+            &mut RenderContext::<_, String>::new(&modifiers, variables, &HashMap::new()),
             &mut rendered,
         )
         .unwrap();
@@ -144,7 +150,7 @@ mod tests {
         let mut rendered = String::new();
 
         tpl.render(
-            &mut RenderContext::new(&modifiers, variables, &HashMap::new()),
+            &mut RenderContext::<_, String>::new(&modifiers, variables, &HashMap::new()),
             &mut rendered,
         )
         .unwrap();
@@ -168,7 +174,7 @@ mod tests {
 
         let mut rendered = String::new();
         tpl.render(
-            &mut RenderContext::new(&modifiers, variables, &HashMap::new()),
+            &mut RenderContext::<_, String>::new(&modifiers, variables, &HashMap::new()),
             &mut rendered,
         )
         .unwrap();
@@ -194,7 +200,7 @@ Baz"#,
 
         let mut rendered = String::new();
         tpl.render(
-            &mut RenderContext::new(&modifiers, variables, &HashMap::new()),
+            &mut RenderContext::<_, String>::new(&modifiers, variables, &HashMap::new()),
             &mut rendered,
         )
         .unwrap();
@@ -213,7 +219,7 @@ Baz"#,
 
         let mut rendered = String::new();
         tpl.render(
-            &mut RenderContext::new(&modifiers, variables, &HashMap::new()),
+            &mut RenderContext::<_, String>::new(&modifiers, variables, &HashMap::new()),
             &mut rendered,
         )
         .unwrap();
@@ -232,7 +238,7 @@ Baz"#,
 
         let mut rendered = String::new();
         tpl.render(
-            &mut RenderContext::new(&modifiers, variables, &HashMap::new()),
+            &mut RenderContext::<_, String>::new(&modifiers, variables, &HashMap::new()),
             &mut rendered,
         )
         .unwrap();
@@ -242,7 +248,7 @@ Baz"#,
         variables.insert("var1".to_owned(), Value::Bool(false));
         let mut rendered = String::new();
         tpl.render(
-            &mut RenderContext::new(&modifiers, variables, &HashMap::new()),
+            &mut RenderContext::<_, String>::new(&modifiers, variables, &HashMap::new()),
             &mut rendered,
         )
         .unwrap();